model name changed and route modified
