+++ application/app/Models/admin/AdminUser.php
@@ -1,0 +1,51 @@
+<?php
+
+namespace App\Models\admin;
+
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Foundation\Auth\User as Authenticatable;
+
+/**
+ * App\Models\AdminUsers
+ *
+ * @property int $id
+ * @property string $name
+ * @property string $email
+ * @property string $password
+ * @property \Illuminate\Support\Carbon|null $created_at
+ * @property \Illuminate\Support\Carbon|null $updated_at
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\AdminUser newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\AdminUser newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\AdminUser query()
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\AdminUser whereCreatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\AdminUser whereEmail($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\AdminUser whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\AdminUser whereName($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\AdminUser wherePassword($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\AdminUser whereUpdatedAt($value)
+ * @mixin \Eloquent
+ */
+class AdminUser extends Authenticatable
+{
+    /**
+     * The attributes that are mass assignable.
+     *
+     * @var array
+     */
+
+    protected $fillable = [
+        'name',
+        'email',
+        'is_owner',
+        'password',
+    ];
+
+    /**
+     * The attributes that should be hidden for arrays.
+     *
+     * @var array
+     */
+    protected $hidden = [
+        'password',
+    ];
+}
