+++ application/app/Models/admin/ProductCategory.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace App\Models\admin;
+
+use Illuminate\Database\Eloquent\Model;
+
+/**
+ * App\Models\ProductCategory
+ *
+ * @property int $id
+ * @property string $name
+ * @property int $order_no
+ * @property \Illuminate\Support\Carbon|null $created_at
+ * @property \Illuminate\Support\Carbon|null $updated_at
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\ProductCategory newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\ProductCategory newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\ProductCategory query()
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\ProductCategory whereCreatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\ProductCategory whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\ProductCategory whereName($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\ProductCategory whereOrderNo($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\admin\ProductCategory whereUpdatedAt($value)
+ * @mixin \Eloquent
+ */
+class ProductCategory extends Model
+{
+    protected $fillable = [
+        'name',
+        'order_no',
+        ];
+
+    protected $hidden = [
+    ];
+
+    protected $casts = [
+    ];
+
+    protected $dates = [
+    ];
+
+    public function scopeFilterLikeName($query, $name)
+    {
+        return $query->where('name','like','%'.$name.'%');
+    }
+}
